+++ src/main/java/com/seminar/easyCookWeb/Converter/EmployeeConverter.java
@@ -2,7 +2,7 @@ package com.seminar.easyCookWeb.Converter;
 
 import com.seminar.easyCookWeb.Entity.User.EmployeeRequest;
 import com.seminar.easyCookWeb.Entity.User.EmployeeResponse;
-import com.seminar.easyCookWeb.Pojo.app_user.Employee;
+import com.seminar.easyCookWeb.Pojo.appUser.Employee;
 import lombok.NoArgsConstructor;
 import org.springframework.beans.BeanUtils;
