+++ src/main/java/com/seminar/easyCookWeb/Mapper/MemberMapper.java
@@ -1,0 +1,4 @@
+package com.seminar.easyCookWeb.Mapper;
+
+public interface MemberMapper {
+}
