+++ src/main/java/com/seminar/easyCookWeb/Pojo/appUser/User.java
@@ -1,54 +1,22 @@
-package com.seminar.easyCookWeb.Pojo.app_user;
+package com.seminar.easyCookWeb.Pojo.appUser;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
-import lombok.ToString;
-import org.hibernate.annotations.Nationalized;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import javax.persistence.*;
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Null;
+import javax.persistence.Transient;
 import java.util.Collection;
 
-@Entity
-@ToString
 @Data
-public class Member extends User implements UserDetails {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Null
-    @Column(name="member_id")
-    private long id;
-
-    @NotEmpty
-    @Column(length = 45, unique = true)
+public abstract class User implements UserDetails {
     private String account;
-
-    @NotEmpty
-    @Column(length = 1024)
-    @JsonIgnore
     private String password;
-
-    @Column(length = 65)
-    @Nationalized
     private String username;
-
-    @Column(columnDefinition = "nvarchar(15)")
-    private String phone;//如果你這裡使用Num大寫，寫入資料庫會變成phone_num
-
-    @Column(columnDefinition = "nvarchar(254)")
+    private String phone;
     private String email;
-
-
-    @NotNull
-    @Enumerated(EnumType.STRING)
-    @Column(length = 8)
     private Role role;
-
 
     @Transient
     @JsonIgnore
