--- src/main/java/com/seminar/easyCookWeb/Pojo/app_user/User.java
@@ -1,55 +1,0 @@
-package com.seminar.easyCookWeb.Pojo.app_user;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Data;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import javax.persistence.Transient;
-import java.util.Collection;
-
-@Data
-public abstract class User implements UserDetails {
-    private String account;
-    private String password;
-    private String username;
-    private String phone;
-    private String email;
-    private Role role;
-
-    @Transient
-    @JsonIgnore
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return AuthorityUtils.createAuthorityList(getRole().getName());
-    }
-
-    @Transient
-    @JsonIgnore
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Transient
-    @JsonIgnore
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Transient
-    @JsonIgnore
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Transient
-    @JsonIgnore
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
