+++ src/main/java/com/seminar/easyCookWeb/Repository/Users/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.seminar.easyCookWeb.Repository.Users;
 
-import com.seminar.easyCookWeb.Pojo.app_user.Member;
+import com.seminar.easyCookWeb.Pojo.appUser.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
