+++ src/main/java/com/seminar/easyCookWeb/Service/User/EmployeeService.java
@@ -3,11 +3,11 @@ package com.seminar.easyCookWeb.Service.User;
 import com.seminar.easyCookWeb.Converter.EmployeeConverter;
 import com.seminar.easyCookWeb.Exception.ConflictException;
 import com.seminar.easyCookWeb.Exception.EntityNotFoundException;
-import com.seminar.easyCookWeb.Pojo.app_user.Role;
+import com.seminar.easyCookWeb.Pojo.appUser.Role;
 import com.seminar.easyCookWeb.Repository.Users.EmployeeRepository;
 import com.seminar.easyCookWeb.Entity.User.EmployeeRequest;
 import com.seminar.easyCookWeb.Entity.User.EmployeeResponse;
-import com.seminar.easyCookWeb.Pojo.app_user.Employee;
+import com.seminar.easyCookWeb.Pojo.appUser.Employee;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
