+++ src/main/java/com/seminar/easyCookWeb/Service/User/MemberService.java
@@ -3,9 +3,9 @@ package com.seminar.easyCookWeb.Service.User;
 import com.seminar.easyCookWeb.Converter.MemberConverter;
 import com.seminar.easyCookWeb.Exception.ConflictException;
 import com.seminar.easyCookWeb.Exception.EntityNotFoundException;
-import com.seminar.easyCookWeb.Pojo.app_user.Role;
+import com.seminar.easyCookWeb.Pojo.appUser.Role;
 import com.seminar.easyCookWeb.Repository.Users.MemberRepository;
-import com.seminar.easyCookWeb.Pojo.app_user.Member;
+import com.seminar.easyCookWeb.Pojo.appUser.Member;
 import com.seminar.easyCookWeb.Entity.User.MemberRequest;
 import com.seminar.easyCookWeb.Entity.User.MemberResponse;
 import lombok.extern.slf4j.Slf4j;
